+++ src/sidebar/js/theme-listener.js
@@ -3,19 +3,19 @@
  */
 
 listenToWindowEvent('start-video', () => {
-    const sidebar = jQuery('body');
+    const sidebar = jQuery('html');
     sidebar.addClass('player-mode');
 });
 
 listenToWindowEvent('play-video', () => {
-    const sidebar = jQuery('body');
+    const sidebar = jQuery('html');
     if (!sidebar.hasClass('player-mode')) {
         sidebar.addClass('player-mode');
     }
 });
 
 listenToWindowEvent('close-video', () => {
-    const sidebar = jQuery('body');
+    const sidebar = jQuery('html');
     sidebar.removeClass('player-mode');
 });